:tractor: Refactored the film develop application into subfiles
