+++ cmd/apps/film-develop-tui/app/doc.go
@@ -1,0 +1,3 @@
+// Package app contains the BubbleTea application model and integration logic.
+// This package depends on ui, state, and types packages.
+package app
