+++ cmd/apps/film-develop-tui/state/doc.go
@@ -1,0 +1,3 @@
+// Package state contains the application state management and state machine logic.
+// This package depends only on the types package.
+package state
