+++ cmd/apps/film-develop-tui/types/doc.go
@@ -1,0 +1,3 @@
+// Package types contains the core domain types for the film development calculator.
+// This package has no external dependencies and contains pure data structures.
+package types
