+++ cmd/apps/film-develop-tui/ui/doc.go
@@ -1,0 +1,3 @@
+// Package ui contains all user interface components including screens and styles.
+// This package depends on types, state, and external UI libraries.
+package ui
