+++ cmd/apps/film-develop-tui/ui/styles.go
@@ -1,0 +1,49 @@
+package ui
+
+import "github.com/charmbracelet/lipgloss"
+
+// Styles for the UI
+var (
+	TitleStyle = lipgloss.NewStyle().
+		Bold(true).
+		Foreground(lipgloss.Color("12")).
+		Align(lipgloss.Center).
+		Width(81)
+
+	BoxStyle = lipgloss.NewStyle().
+		Border(lipgloss.RoundedBorder()).
+		BorderForeground(lipgloss.Color("8")).
+		Padding(1).
+		Width(79)
+
+	MainBoxStyle = lipgloss.NewStyle().
+		Border(lipgloss.RoundedBorder()).
+		BorderForeground(lipgloss.Color("8")).
+		Padding(1).
+		Width(79)
+
+	SectionStyle = lipgloss.NewStyle().
+		Bold(true).
+		Foreground(lipgloss.Color("11"))
+
+	HighlightStyle = lipgloss.NewStyle().
+		Bold(true).
+		Foreground(lipgloss.Color("10"))
+
+	DimStyle = lipgloss.NewStyle().
+		Foreground(lipgloss.Color("8"))
+
+	ErrorStyle = lipgloss.NewStyle().
+		Bold(true).
+		Foreground(lipgloss.Color("9"))
+
+	ActionStyle = lipgloss.NewStyle().
+		Bold(true).
+		Foreground(lipgloss.Color("14"))
+
+	ActionsOnlyStyle = lipgloss.NewStyle().
+		Bold(true).
+		Foreground(lipgloss.Color("14")).
+		Align(lipgloss.Center).
+		Width(79)
+)
